Use references for links to user profiles
Those links will be repeated every day and we don't need to keep n
copies of the same link around. Instead of copying the link every day we
can now use references.

If you want to add an entry with an unknown profile, please also use a
reference by first adding a link to your profile in the reference list
at the bottom (the list is sorted alphabetically).